@@ -22,12 +22,14 @@ Gem::Specification.new do |s|
   s.add_dependency 'capybara', '~> 3.0'
 
   s.add_development_dependency 'rspec-given', '~> 3.5.0'
-  s.add_development_dependency 'capybara-webkit', '~> 1.7.0'
-  s.add_development_dependency 'poltergeist', '~> 1.17.0'
+  s.add_development_dependency 'selenium-webdriver'
+  s.add_development_dependency 'webdrivers'
+  s.add_development_dependency 'cuprite'
   s.add_development_dependency 'cucumber', '~> 1.3.0'
   s.add_development_dependency 'pry'
   s.add_development_dependency 'pry-nav'
+  s.add_development_dependency 'puma'
   s.add_development_dependency 'sinatra'
-  s.add_development_dependency 'rails', '~> 4.2.9'
+  s.add_development_dependency 'rails', '~> 5.2.0'
   s.add_development_dependency 'simplecov'
 end
