@@ -1,5 +1,7 @@
-require 'capybara/poltergeist'
+require 'capybara/cuprite'
+require 'webdrivers'
 
 Capybara.default_max_wait_time = 2
 Capybara.app = CapybaraUIApp
-Capybara.javascript_driver = :poltergeist
+Capybara.javascript_driver = :selenium_chrome_headless
+Capybara.current_driver = :selenium_chrome_headless
